Revert "Web and db wrapper"
