+++ .gitignore
@@ -3,5 +3,3 @@
 build/**/*
 .idea/**
 *.iml
-maths.db
-report.html
