+++ robotone.cabal
@@ -16,21 +16,17 @@ cabal-version:       >=1.10
  
 library
   hs-source-dirs:      src
-  exposed-modules:     Writeup Types TidyingMoves TexBase Tex TestData TestData2
+  exposed-modules:     Writeup Types TidyingMoves TexBase Tex TestData TestData2 TestData3
                        Suspension RobotM Rename Printing Parser Move Match Library
-                       Expansion DeletionMoves ApplyingMoves Server
-  build-depends:       base >= 4.7 && < 5, containers, mtl, parsec, logict, QuickCheck, transformers, warp, aeson,
-   wai-make-assets, servant, wai, servant-server, text, time, servant-blaze, blaze-html, HDBC, HDBC-sqlite3, json, aeson,
-   either, regex-compat
+                       Expansion DeletionMoves ApplyingMoves
+  build-depends:       base >= 4.7 && < 5, containers, mtl, parsec, logict, QuickCheck, transformers
   default-language:    Haskell2010
 
 
 executable robotone-exe
-  hs-source-dirs:     app, src
+  hs-source-dirs:     app
   main-is:            Main.hs
-  build-depends:      base, robotone, containers, mtl, transformers, logict, warp, aeson, wai-make-assets,
-   servant, wai, servant-server, text, time, servant-blaze, blaze-html, HDBC, HDBC-sqlite3, json, aeson, either, regex-compat, parsec
-
+  build-depends:      base, robotone, containers, mtl, transformers, logict
   default-language:   Haskell2010
