+++ run.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 wd=$(dirname "$0")
 
-xelatex=$(which tex)
+xelatex=$(which xelatex)
 if [$xelatex == ""]
 then
    xelatex=$(which pdflatex)
