+++ src/Suspension.hs
@@ -12,7 +12,6 @@ import Types
 import Move
 import Tex
 import RobotM
-import WriteupBase
 import Writeup
 
 ----------------------------------------------------------------------------------------------------
