+++ stack.yaml
@@ -2,18 +2,18 @@
 # For more information, see: http://docs.haskellstack.org/en/stable/yaml_configuration/
 
 # Specifies the GHC version and set of packages available (e.g., lts-3.5, nightly-2015-09-21, ghc-7.10.2)
-resolver: ghc-8.4.3 #lts-11.4 #ghc-8.4.1
+resolver: ghc-8.4.1
 
 # Local packages, usually specified by relative directory name
 packages:
 - '.'
-
 # Packages to be pulled from upstream that are not in the resolver (e.g., acme-missiles-0.3)
-extra-deps: [HDBC-2.4.0.2,HDBC-sqlite3-2.3.3.1,QuickCheck-2.11.3,aeson-1.4.0.0,blaze-html-0.9.1.1,either-5,json-0.9.2,logict-0.6.0.2,regex-compat-0.95.1,servant-0.14,servant-blaze-0.8,servant-server-0.14,wai-3.2.1.2,wai-make-assets-0.2,warp-3.2.22,async-2.2.1,attoparsec-0.13.2.2,auto-update-0.1.4,base-compat-0.10.1,base64-bytestring-1.0.0.1,bifunctors-5.5.2,blaze-builder-0.4.1.0,blaze-markup-0.8.2.1,bsb-http-chunked-0.0.0.2,cabal-doctest-1.0.6,case-insensitive-1.2.0.11,convertible-1.1.1.0,data-default-0.7.1.1,dlist-0.8.0.4,exceptions-0.10.0,getopt-generics-0.13.0.2,hashable-1.2.7.0,http-api-data-0.3.8.1,http-date-0.0.7,http-media-0.7.1.2,http-types-0.12.1,http2-1.6.3,iproute-1.7.5,mmorph-1.1.2,monad-control-1.0.2.3,natural-transformation-0.4,network-2.7.0.1,network-uri-2.6.1.0,old-time-1.1.0.3,profunctors-5.2.2,random-1.1,regex-base-0.93.2,regex-posix-0.95.2,resourcet-1.2.1,scientific-0.3.6.2,semigroupoids-5.2.2,semigroups-0.18.4,shake-0.16.4,simple-sendfile-0.2.27,singleton-bool-0.1.4,streaming-commons-0.2.0.0,string-conversions-0.4.0.1,syb-0.7,tagged-0.8.5,tf-random-0.5,th-abstraction-0.2.7.0,time-locale-compat-0.1.1.4,transformers-base-0.4.5.2,transformers-compat-0.6.2,unix-compat-0.5.0.1,unordered-containers-0.2.9.0,utf8-string-1.0.1.1,uuid-types-1.0.3,vault-0.3.1.1,vector-0.12.0.1,wai-app-static-3.1.6.2,word8-0.1.3,appar-0.1.4,attoparsec-iso8601-1.0.0.0,base-orphans-0.7,byteorder-1.0.4,bytestring-builder-0.10.8.1.0,comonad-5.0.3,contravariant-1.4.1,cryptonite-0.25,data-default-class-0.1.2.0,data-default-instances-containers-0.0.1,data-default-instances-dlist-0.0.1,data-default-instances-old-locale-0.0.1,distributive-0.5.3,extra-1.6.8,file-embed-0.0.10.1,generics-sop-0.3.2.0,integer-logarithms-1.0.2.1,js-flot-0.8.3,js-jquery-3.3.1,memory-0.14.16,mime-types-0.1.0.8,old-locale-1.0.0.7,optparse-applicative-0.14.2.0,primitive-0.6.4.0,psqueues-0.2.7.0,unliftio-core-0.1.1.0,uri-bytestring-0.3.2.0,wai-extra-3.0.22.1,zlib-0.6.2,StateVar-1.1.1.0,ansi-terminal-0.8.0.4,ansi-wl-pprint-0.6.8.2,basement-0.0.7,clock-0.7.2,cookie-0.4.4,fast-logger-2.4.11,foundation-0.0.20,lifted-base-0.2.3.12,stringsearch-0.3.6.6,th-lift-instances-0.1.11,void-0.7.2,wai-logger-2.3.2,colour-2.3.4,easy-file-0.2.2,th-lift-0.7.10,unix-time-0.3.8,hspec-2.5.1,HUnit-1.6.0.0,call-stack-0.1.0,hspec-core-2.5.1,hspec-discover-2.5.1,hspec-expectations-0.8.2,quickcheck-io-0.2.0,setenv-0.1.1.3] #[wai-make-assets-0.2]
+extra-deps: [hspec-2.5.0, QuickCheck-2.11.3, logict-0.6.0.2, random-1.1, tf-random-0.5, primitive-0.6.3.0,
+HUnit-1.6.0.0, call-stack-0.1.0, hspec-core-2.5.0, hspec-discover-2.5.0, hspec-expectations-0.8.2, ansi-terminal-0.8.0.2,
+clock-0.7.2, quickcheck-io-0.2.0, setenv-0.1.1.3, colour-2.3.4]
 
 # Override default flag values for local packages and extra-deps
-flags:
-
+flags: {}
 
 # Extra package databases containing global packages
 extra-package-dbs: []
@@ -35,4 +35,3 @@ extra-package-dbs: []
 
 # Allow a newer minor version of GHC than the snapshot specifies
 compiler-check: newer-minor
-allow-newer: true
